[refactor] interfaces 경로 제거
