+++ src/main/java/com/hhplus/concert/api/reservation/presentation/dto/request/ReservationReqDTO.java
@@ -1,4 +1,4 @@
-package com.hhplus.concert.api.reservation.interfaces.presentation.dto.request;
+package com.hhplus.concert.api.reservation.presentation.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
