+++ src/main/java/com/hhplus/concert/api/reservation/presentation/dto/response/PaymentResDTO.java
@@ -1,4 +1,4 @@
-package com.hhplus.concert.api.reservation.interfaces.presentation.dto.response;
+package com.hhplus.concert.api.reservation.presentation.dto.response;
 
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
